Implement suggestions
