+++ packages/smooth_app/lib/helpers/image_field_extension.dart
@@ -85,6 +85,22 @@ extension ImageFieldSmoothieExtension on ImageField {
         ImageField.OTHER => appLocalizations.take_more_photo_button_label,
       };
 
+  String getPictureAccessibilityLabel(
+    final AppLocalizations appLocalizations,
+  ) =>
+      switch (this) {
+        ImageField.FRONT =>
+          appLocalizations.product_image_front_accessibility_label,
+        ImageField.INGREDIENTS =>
+          appLocalizations.product_image_ingredients_accessibility_label,
+        ImageField.NUTRITION =>
+          appLocalizations.product_image_nutrition_accessibility_label,
+        ImageField.PACKAGING =>
+          appLocalizations.product_image_packaging_accessibility_label,
+        ImageField.OTHER =>
+          appLocalizations.product_image_other_accessibility_label,
+      };
+
   Widget getPhotoButton(
     final BuildContext context,
     final Product product,
